Movo reusable components to a separate folder
